On branch tuple-rest-sandwich-work  Changes to be committed: 	modified:   src/compiler/checker.ts 	modified:   tmp.cph.d/FIX_tuple-rest.md 	renamed:    tests/cases/compiler/_tlt.ts -> tmp.cph.d/_tlt.ts
Completed unification of old and new code.

These functions exable the compatibility:
 getContextualTypeForElementExpression_altTypeAndFlag
 getContextualTypeForElementExpression_altLength
 getContextualTypeForElementExpression_altFlags

Still there is no real length for some types which are not isTupleType.
See getContextualTypeForElementExpression_altFlags for how that works now
by returning a large number instead of a real length.

However, that won't in case of bidirectional search.
So it's just lucky that the baseline tests don't contain a comnbination
of a type that has no defined length and the need for a bidirection search.