On branch condalias-49075-work-2
While testing _cax-a2, which has a while loop
the correct results are being returned at each top level `checkExpression`, according to the log files,
but
- A. the final result written to the xxxx.types file doesn't reflect that and contains errors
- B. the final result written to the xxxx.errors.txt file DIES reflect the top level `checkExpression` results.

I think what may be happening is that the `sharedFlowNodes` cache written/read in `getTypeAtFlowNode` is somehow polluted
with an intermediate value, and that value is somehow read back for case A above.  Maybe.
That works OK for the existing algorithm because the types are generally over-wide, but fails for the new aliasing algo
because the types are more narrow.

Plan:  Try a different cache strategy, caching the results at the  `checkExpression` level, because those appear to be correct anyway.
Read the cache
(1) At `getFlowTypeOfReference`
(2) Maybe also at `getTypeAtFlowNode` - maybe.