On branch condalias-49075-work-2
The problem was not with sharedCashed but with typeChecker top level scoped flowTypeCache,
which is also used in getTypeAtFlowLoopLabel, as well as numerous other places.
However, double checking shows that the return values of checkExpression,
while not inside getFlowTypeOfReference, indicates those values are reliable.
(Presumably those are the same as what is being returned from top level getFlowTypeOfReference.)
So we'll make a cache for those, and make sure that getTypeAtLocation checks that cache first.