On branch condalias-49075-work-2
Not a stable version, but want to peg it anyway.
This is worth solving.  But I should roll back some changes made for caching when `isOriginalCall` is false:  `altTypeCache`.
But doing that will result in hanging on `binaryArithmeticControlFlowGraphNotTooLarge`.
To avoid that it seems necessary to roll back all caching changes - get rid of `typeCache` and go back to using only the existing `cache`.
Then check complete test suite.
Once that is stable, then investigate this problem further.