On branch condalias-49075-work-2
Now we are back to passing all runtests, but the bug fix no longer works inside loops.
I.e., _cax-a4 is working, but not _cax-a1 or _cax-a2.
The problem is probably that `getTypeOfExpression` again reads cached values
```
            if (/* flowTypeQueryState.disable && */ node.flags & NodeFlags.TypeCached && flowTypeCache) {
                const cachedType = flowTypeCache[getNodeId(node)];
                if (cachedType) {
                    return cachedType;
                }
            }
```
despite the fact that checkExpression itself is returning correct values.
Next:  Try overwriting `flowTypeCache[getNodeId(node)]` with the return value of `getFlowTypeOfReference` when the call is not recursive.