On branch condalias-49075-work-2
I have a hunch the problem is related to `isReachableFlowNode` which traverses all the flowNodes at every
`getTypeAtFlowAssignment`, and thus has to traverse the FlowJoin's as well.  That's problematic as the
mapping of a FlowJoin antecedent varies by case.

For one thig!  The `isReachableFlowNode` processing is O(N^2) behavior. Dame desu.

To get around this, let us try a different strategy:
First continue processing from the FlowJoin.antecedent, let the result be type0.
Then call the aliasFlow with join mapping to FlowStart, so that isReachableFlowNode will stay happy.